% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaintegerMutation.R
\name{gaintegerMutation}
\alias{gaintegerMutation}
\title{Integer mutation for genetic algorithm}
\usage{
gaintegerMutation(object, parent, ...)
}
\arguments{
\item{object}{GA object containing algorithm parameters.}

\item{parent}{Integer index of the parent individual to mutate.}

\item{...}{Additional arguments (not used).}
}
\value{
Numeric vector representing the mutated individual.
}
\description{
Performs mutation on a genetic algorithm individual by randomly
changing cluster assignments with a specified probability.
}
