% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaintegerOnePointCrossover.R
\name{gaintegerOnePointCrossover}
\alias{gaintegerOnePointCrossover}
\title{One-point crossover for genetic algorithm with integer encoding}
\usage{
gaintegerOnePointCrossover(object, parents, ...)
}
\arguments{
\item{object}{GA object containing algorithm parameters.}

\item{parents}{Integer vector of length 2 containing indices of parent individuals.}

\item{...}{Additional arguments (not used).}
}
\value{
List containing:
\describe{
\item{children}{Matrix with two rows representing the offspring}
\item{fitness}{Vector of NA values (fitness will be calculated later)}
}
}
\description{
Performs one-point crossover between two parent individuals in the
genetic algorithm, exchanging genetic material at a single randomly selected point.
}
