% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twomodeToy.R
\docType{data}
\name{twomodeToy}
\alias{twomodeToy}
\title{Toy matrix with one multiplicative and one additive bicluster}
\format{
A numeric matrix of dimension \eqn{12 \times 9} with a 2 x 3 two-mode cluster structure
}
\usage{
data(twomodeToy)
}
\description{
A small 12×9 matrix with a 2 x 3 two-mode cluster structure to demonstrate
twomodeclusteringGA in a controlled setting.
}
\examples{
data("twomodeToy")
str(twomodeToy)
image(t(twomodeToy))
}
\keyword{datasets}
